finish test refactoring
